# Optimize `regex_replace` for scalar patterns